pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atdd-orchestrator"
description = "Orquestador ATDD — arquitectura hexagonal"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.4",
    "redis>=5.0",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-celery",
]

[tool.hatch.build.targets.wheel]
packages = ["atdd_orchestrator"]
